renamed db to database
