+++ app/src/main/java/com/example/sbma_project/database/Run.kt
@@ -1,4 +1,4 @@
-package com.example.sbma_project.DB
+package com.example.sbma_project.database
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
